refactor(process): migrate processcheck component to v2
