+++ comp/otelcol/logsagentpipeline/doc.go
@@ -1,0 +1,7 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+// Package logsagentpipeline provides the logs agent pipeline component.
+package logsagentpipeline
