+++ comp/otelcol/logsagentpipeline/impl/statusimpl.go
