+++ comp/process/processcheck/def/component.go
