+++ comp/process/processcheck/impl/check.go
@@ -7,8 +7,8 @@
 package processcheckimpl
 
 import (
+	compdef "github.com/DataDog/datadog-agent/comp/def"
 	"github.com/DataDog/datadog-go/v5/statsd"
-	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
@@ -16,17 +16,10 @@ import (
 	taggerdef "github.com/DataDog/datadog-agent/comp/core/tagger/def"
 	workloadmeta "github.com/DataDog/datadog-agent/comp/core/workloadmeta/def"
 	gpusubscriber "github.com/DataDog/datadog-agent/comp/process/gpusubscriber/def"
-	"github.com/DataDog/datadog-agent/comp/process/processcheck"
+	processcheck "github.com/DataDog/datadog-agent/comp/process/processcheck/def"
 	"github.com/DataDog/datadog-agent/comp/process/types"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newCheck))
-}
 
 var _ types.CheckComponent = (*check)(nil)
 
@@ -35,7 +28,7 @@ type check struct {
 }
 
 type dependencies struct {
-	fx.In
+	compdef.In
 
 	Config        config.Component
 	Sysconfig     sysprobeconfig.Component
@@ -46,18 +39,19 @@ type dependencies struct {
 	Tagger        taggerdef.Component
 }
 
-type result struct {
-	fx.Out
+type Provides struct {
+	compdef.Out
 
 	Check     types.ProvidesCheck
 	Component processcheck.Component
 }
 
-func newCheck(deps dependencies) result {
+// NewCheck creates a new processcheck component.
+func NewCheck(deps dependencies) Provides {
 	c := &check{
 		processCheck: checks.NewProcessCheck(deps.Config, deps.Sysconfig, deps.WMmeta, deps.GpuSubscriber, deps.Statsd, deps.IPC.GetTLSServerConfig(), deps.Tagger),
 	}
-	return result{
+	return Provides{
 		Check: types.ProvidesCheck{
 			CheckComponent: c,
 		},
