+++ comp/process/runner/runnerimpl/runner_test.go
@@ -25,7 +25,7 @@ import (
 	workloadmetafx "github.com/DataDog/datadog-agent/comp/core/workloadmeta/fx"
 	"github.com/DataDog/datadog-agent/comp/process/containercheck/containercheckimpl"
 	hostinfomock "github.com/DataDog/datadog-agent/comp/process/hostinfo/mock"
-	"github.com/DataDog/datadog-agent/comp/process/processcheck/processcheckimpl"
+	processcheckimpl "github.com/DataDog/datadog-agent/comp/process/processcheck/impl"
 	"github.com/DataDog/datadog-agent/comp/process/runner"
 	"github.com/DataDog/datadog-agent/comp/process/submitter/submitterimpl"
 	"github.com/DataDog/datadog-agent/comp/process/types"
@@ -118,7 +118,7 @@ func createDeps(t *testing.T, confOverrides map[string]interface{}, options ...f
 		hostinfomock.MockModule(),
 
 		// Checks
-		processcheckimpl.MockModule(),
+		fx.Provide(processcheckimpl.NewMock),
 		containercheckimpl.MockModule(),
 
 		fx.Provide(func(t testing.TB) log.Component { return logmock.New(t) }),
