+++ go.work
@@ -62,7 +62,6 @@ use (
 	comp/otelcol/ddprofilingextension/def
 	comp/otelcol/ddprofilingextension/impl
 	comp/otelcol/logsagentpipeline
-	comp/otelcol/logsagentpipeline/logsagentpipelineimpl
 	comp/otelcol/otlp/components/exporter/datadogexporter
 	comp/otelcol/otlp/components/exporter/logsagentexporter
 	comp/otelcol/otlp/components/exporter/serializerexporter
