+++ modules.yml
@@ -114,8 +114,6 @@ modules:
     used_by_otel: true
   comp/otelcol/logsagentpipeline:
     used_by_otel: true
-  comp/otelcol/logsagentpipeline/logsagentpipelineimpl:
-    used_by_otel: true
   comp/otelcol/otlp/components/exporter/datadogexporter:
     used_by_otel: true
   comp/otelcol/otlp/components/exporter/logsagentexporter:
